code/requirements.txt
certifi==2025.1.31
influxdb-client==1.48.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
